requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pillow==10.2.0
transformers==4.37.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
